🏗️ Architect: Modularized generic adapter vendor data
Refactored the 14,676-line `apts/opticalequipment/adapter/vendors/generic.py` 'God Object' into a modular package `apts/opticalequipment/adapter/vendors/generic/`.

- Split `GenericAdapter` class and its database into `adapters.py`.
- Split `GenericSpacer` class and its database into `spacers.py`.
- Maintained public API and dynamic loading compatibility via `__init__.py`.
- Updated relative imports to account for new directory nesting.

This change significantly improves maintainability of the large equipment database by breaking down the largest single file in the repository into logical components without breaking existing functionality.
