+++ apts/opticalequipment/adapter/vendors/generic/__init__.py
@@ -1,0 +1,2 @@
+from .adapters import GenericAdapter
+from .spacers import GenericSpacer
