+++ apts/opticalequipment/adapter/vendors/generic/spacers.py
@@ -1,0 +1,1738 @@
+from ...base import Spacer
+class GenericSpacer(Spacer):
+    _DATABASE = {
+        "Generic_M42_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M72_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M72 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 13,
+            "tside_thread": "M72",
+            "tside_gender": "",
+            "cside_thread": "M72",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M72_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M72 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 13,
+            "tside_thread": "M72",
+            "tside_gender": "",
+            "cside_thread": "M72",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M82_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M82 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 16,
+            "tside_thread": "M82",
+            "tside_gender": "",
+            "cside_thread": "M82",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_3mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.3mm",
+            "type": "type_spacer",
+            "optical_length": 0.3,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_3mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.3mm",
+            "type": "type_spacer",
+            "optical_length": 0.3,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_3mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.3mm",
+            "type": "type_spacer",
+            "optical_length": 0.3,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M92_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M92 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 19,
+            "tside_thread": "M92",
+            "tside_gender": "",
+            "cside_thread": "M92",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M56_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M56 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 9,
+            "tside_thread": "M56",
+            "tside_gender": "",
+            "cside_thread": "M56",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M56_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M56 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 9,
+            "tside_thread": "M56",
+            "tside_gender": "",
+            "cside_thread": "M56",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M63_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M63 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 11,
+            "tside_thread": "M63",
+            "tside_gender": "",
+            "cside_thread": "M63",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M63_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M63 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 11,
+            "tside_thread": "M63",
+            "tside_gender": "",
+            "cside_thread": "M63",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M84_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M84 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 15,
+            "tside_thread": "M84",
+            "tside_gender": "",
+            "cside_thread": "M84",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_SC_Schmidt_Cassegrain_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "SC (Schmidt-Cassegrain) Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 21,
+            "tside_thread": "SC (Schmidt-Cassegrain)",
+            "tside_gender": "",
+            "cside_thread": "SC (Schmidt-Cassegrain)",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_SC_Schmidt_Cassegrain_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "SC (Schmidt-Cassegrain) Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 21,
+            "tside_thread": "SC (Schmidt-Cassegrain)",
+            "tside_gender": "",
+            "cside_thread": "SC (Schmidt-Cassegrain)",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_1mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.1mm",
+            "type": "type_spacer",
+            "optical_length": 0.1,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_15mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.15mm",
+            "type": "type_spacer",
+            "optical_length": 0.15,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_2mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.2mm",
+            "type": "type_spacer",
+            "optical_length": 0.2,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_25mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.25mm",
+            "type": "type_spacer",
+            "optical_length": 0.25,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_35mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.35mm",
+            "type": "type_spacer",
+            "optical_length": 0.35,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_6mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.6mm",
+            "type": "type_spacer",
+            "optical_length": 0.6,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_75mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.75mm",
+            "type": "type_spacer",
+            "optical_length": 0.75,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_0_8mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 0.8mm",
+            "type": "type_spacer",
+            "optical_length": 0.8,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_1mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.1mm",
+            "type": "type_spacer",
+            "optical_length": 0.1,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_15mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.15mm",
+            "type": "type_spacer",
+            "optical_length": 0.15,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_2mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.2mm",
+            "type": "type_spacer",
+            "optical_length": 0.2,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_25mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.25mm",
+            "type": "type_spacer",
+            "optical_length": 0.25,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_35mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.35mm",
+            "type": "type_spacer",
+            "optical_length": 0.35,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_6mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.6mm",
+            "type": "type_spacer",
+            "optical_length": 0.6,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_75mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.75mm",
+            "type": "type_spacer",
+            "optical_length": 0.75,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_0_8mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 0.8mm",
+            "type": "type_spacer",
+            "optical_length": 0.8,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_1mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.1mm",
+            "type": "type_spacer",
+            "optical_length": 0.1,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_15mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.15mm",
+            "type": "type_spacer",
+            "optical_length": 0.15,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_2mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.2mm",
+            "type": "type_spacer",
+            "optical_length": 0.2,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_25mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.25mm",
+            "type": "type_spacer",
+            "optical_length": 0.25,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_35mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.35mm",
+            "type": "type_spacer",
+            "optical_length": 0.35,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_6mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.6mm",
+            "type": "type_spacer",
+            "optical_length": 0.6,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_75mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.75mm",
+            "type": "type_spacer",
+            "optical_length": 0.75,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_0_8mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 0.8mm",
+            "type": "type_spacer",
+            "optical_length": 0.8,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_0_75mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 0.75mm",
+            "type": "type_spacer",
+            "optical_length": 0.75,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M72_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M72 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 13,
+            "tside_thread": "M72",
+            "tside_gender": "",
+            "cside_thread": "M72",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M72_Spacer_0_75mm": {
+            "brand": "Generic",
+            "name": "M72 Spacer 0.75mm",
+            "type": "type_spacer",
+            "optical_length": 0.75,
+            "mass": 13,
+            "tside_thread": "M72",
+            "tside_gender": "",
+            "cside_thread": "M72",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M72_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M72 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 13,
+            "tside_thread": "M72",
+            "tside_gender": "",
+            "cside_thread": "M72",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M82_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M82 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 16,
+            "tside_thread": "M82",
+            "tside_gender": "",
+            "cside_thread": "M82",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M82_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M82 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 16,
+            "tside_thread": "M82",
+            "tside_gender": "",
+            "cside_thread": "M82",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M82_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M82 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 16,
+            "tside_thread": "M82",
+            "tside_gender": "",
+            "cside_thread": "M82",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M84_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M84 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 15,
+            "tside_thread": "M84",
+            "tside_gender": "",
+            "cside_thread": "M84",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M92_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M92 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 17,
+            "tside_thread": "M92",
+            "tside_gender": "",
+            "cside_thread": "M92",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M117_Spacer_1mm": {
+            "brand": "Generic",
+            "name": "M117 Spacer 1mm",
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 23,
+            "tside_thread": "M117",
+            "tside_gender": "",
+            "cside_thread": "M117",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M117_Spacer_2mm": {
+            "brand": "Generic",
+            "name": "M117 Spacer 2mm",
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 23,
+            "tside_thread": "M117",
+            "tside_gender": "",
+            "cside_thread": "M117",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_1_1mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 1.1mm",
+            "type": "type_spacer",
+            "optical_length": 1.1,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_1_3mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 1.3mm",
+            "type": "type_spacer",
+            "optical_length": 1.3,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_1_7mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 1.7mm",
+            "type": "type_spacer",
+            "optical_length": 1.7,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M42_Spacer_1_9mm": {
+            "brand": "Generic",
+            "name": "M42 Spacer 1.9mm",
+            "type": "type_spacer",
+            "optical_length": 1.9,
+            "mass": 5,
+            "tside_thread": "M42",
+            "tside_gender": "",
+            "cside_thread": "M42",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_1_1mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 1.1mm",
+            "type": "type_spacer",
+            "optical_length": 1.1,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_1_3mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 1.3mm",
+            "type": "type_spacer",
+            "optical_length": 1.3,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_1_7mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 1.7mm",
+            "type": "type_spacer",
+            "optical_length": 1.7,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M48_Spacer_1_9mm": {
+            "brand": "Generic",
+            "name": "M48 Spacer 1.9mm",
+            "type": "type_spacer",
+            "optical_length": 1.9,
+            "mass": 7,
+            "tside_thread": "M48",
+            "tside_gender": "",
+            "cside_thread": "M48",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_1_1mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 1.1mm",
+            "type": "type_spacer",
+            "optical_length": 1.1,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_1_3mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 1.3mm",
+            "type": "type_spacer",
+            "optical_length": 1.3,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_1_7mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 1.7mm",
+            "type": "type_spacer",
+            "optical_length": 1.7,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M54_Spacer_1_9mm": {
+            "brand": "Generic",
+            "name": "M54 Spacer 1.9mm",
+            "type": "type_spacer",
+            "optical_length": 1.9,
+            "mass": 9,
+            "tside_thread": "M54",
+            "tside_gender": "",
+            "cside_thread": "M54",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1_1mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1.1mm",
+            "type": "type_spacer",
+            "optical_length": 1.1,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1_3mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1.3mm",
+            "type": "type_spacer",
+            "optical_length": 1.3,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1_7mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1.7mm",
+            "type": "type_spacer",
+            "optical_length": 1.7,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M68_Spacer_1_9mm": {
+            "brand": "Generic",
+            "name": "M68 Spacer 1.9mm",
+            "type": "type_spacer",
+            "optical_length": 1.9,
+            "mass": 11,
+            "tside_thread": "M68",
+            "tside_gender": "",
+            "cside_thread": "M68",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M56_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M56 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 9,
+            "tside_thread": "M56",
+            "tside_gender": "",
+            "cside_thread": "M56",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M56_Spacer_0_7mm": {
+            "brand": "Generic",
+            "name": "M56 Spacer 0.7mm",
+            "type": "type_spacer",
+            "optical_length": 0.7,
+            "mass": 9,
+            "tside_thread": "M56",
+            "tside_gender": "",
+            "cside_thread": "M56",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M56_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M56 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 9,
+            "tside_thread": "M56",
+            "tside_gender": "",
+            "cside_thread": "M56",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M63_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M63 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 11,
+            "tside_thread": "M63",
+            "tside_gender": "",
+            "cside_thread": "M63",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M63_Spacer_0_7mm": {
+            "brand": "Generic",
+            "name": "M63 Spacer 0.7mm",
+            "type": "type_spacer",
+            "optical_length": 0.7,
+            "mass": 11,
+            "tside_thread": "M63",
+            "tside_gender": "",
+            "cside_thread": "M63",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M63_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M63 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 11,
+            "tside_thread": "M63",
+            "tside_gender": "",
+            "cside_thread": "M63",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_SC_Schmidt_Cassegrain_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "SC (Schmidt-Cassegrain) Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 21,
+            "tside_thread": "SC (Schmidt-Cassegrain)",
+            "tside_gender": "",
+            "cside_thread": "SC (Schmidt-Cassegrain)",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_SC_Schmidt_Cassegrain_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "SC (Schmidt-Cassegrain) Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 21,
+            "tside_thread": "SC (Schmidt-Cassegrain)",
+            "tside_gender": "",
+            "cside_thread": "SC (Schmidt-Cassegrain)",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_1_25_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": '1.25" Spacer 0.5mm',
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 4,
+            "tside_thread": '1.25"',
+            "tside_gender": "Female",
+            "cside_thread": '1.25"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_1_25_Spacer_1mm": {
+            "brand": "Generic",
+            "name": '1.25" Spacer 1mm',
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 4,
+            "tside_thread": '1.25"',
+            "tside_gender": "Female",
+            "cside_thread": '1.25"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_1_25_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": '1.25" Spacer 1.5mm',
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 4,
+            "tside_thread": '1.25"',
+            "tside_gender": "Female",
+            "cside_thread": '1.25"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_1_25_Spacer_2mm": {
+            "brand": "Generic",
+            "name": '1.25" Spacer 2mm',
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 4,
+            "tside_thread": '1.25"',
+            "tside_gender": "Female",
+            "cside_thread": '1.25"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_2_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": '2" Spacer 0.5mm',
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 9,
+            "tside_thread": '2"',
+            "tside_gender": "Female",
+            "cside_thread": '2"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_2_Spacer_1mm": {
+            "brand": "Generic",
+            "name": '2" Spacer 1mm',
+            "type": "type_spacer",
+            "optical_length": 1,
+            "mass": 9,
+            "tside_thread": '2"',
+            "tside_gender": "Female",
+            "cside_thread": '2"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_2_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": '2" Spacer 1.5mm',
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 9,
+            "tside_thread": '2"',
+            "tside_gender": "Female",
+            "cside_thread": '2"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_2_Spacer_2mm": {
+            "brand": "Generic",
+            "name": '2" Spacer 2mm',
+            "type": "type_spacer",
+            "optical_length": 2,
+            "mass": 9,
+            "tside_thread": '2"',
+            "tside_gender": "Female",
+            "cside_thread": '2"',
+            "cside_gender": "Male",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M92_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M92 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 17,
+            "tside_thread": "M92",
+            "tside_gender": "",
+            "cside_thread": "M92",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M92_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M92 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 17,
+            "tside_thread": "M92",
+            "tside_gender": "",
+            "cside_thread": "M92",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M117_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M117 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 23,
+            "tside_thread": "M117",
+            "tside_gender": "",
+            "cside_thread": "M117",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M117_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M117 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 23,
+            "tside_thread": "M117",
+            "tside_gender": "",
+            "cside_thread": "M117",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M84_Spacer_0_5mm": {
+            "brand": "Generic",
+            "name": "M84 Spacer 0.5mm",
+            "type": "type_spacer",
+            "optical_length": 0.5,
+            "mass": 15,
+            "tside_thread": "M84",
+            "tside_gender": "",
+            "cside_thread": "M84",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_M84_Spacer_1_5mm": {
+            "brand": "Generic",
+            "name": "M84 Spacer 1.5mm",
+            "type": "type_spacer",
+            "optical_length": 1.5,
+            "mass": 15,
+            "tside_thread": "M84",
+            "tside_gender": "",
+            "cside_thread": "M84",
+            "cside_gender": "",
+            "reversible": True,
+            "bf_role": "",
+        },
+        "Generic_T_Ring_Canon_EOS": {
+            "brand": "Generic",
+            "name": "T-Ring for Canon EOS",
+            "type": "type_adapter",
+            "optical_length": 10,
+            "mass": 40,
+            "tside_thread": "T2",
+            "tside_gender": "Female",
+            "cside_thread": "EOS",
+            "cside_gender": "Male",
+            "reversible": False,
+            "bf_role": "",
+        },
+        "Generic_M42_to_T2_Adapter": {
+            "brand": "Generic",
+            "name": "M42 to T2 Adapter",
+            "type": "type_adapter",
+            "optical_length": 0,
+            "mass": 20,
+            "tside_thread": "T2",
+            "tside_gender": "Female",
+            "cside_thread": "M42",
+            "cside_gender": "Male",
+            "reversible": False,
+            "bf_role": "",
+        },
+    }
+
+    @classmethod
+    def Generic_M42_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M72_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M72_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M72_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M72_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M82_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M82_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_3mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_3mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_3mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M92_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M92_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M56_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M56_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M56_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M56_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M63_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M63_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M63_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M63_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M84_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M84_Spacer_2mm"])
+
+    @classmethod
+    def Generic_SC_Schmidt_Cassegrain_Spacer_1mm(cls):
+        return cls.from_database(
+            cls._DATABASE["Generic_SC_Schmidt_Cassegrain_Spacer_1mm"]
+        )
+
+    @classmethod
+    def Generic_SC_Schmidt_Cassegrain_Spacer_2mm(cls):
+        return cls.from_database(
+            cls._DATABASE["Generic_SC_Schmidt_Cassegrain_Spacer_2mm"]
+        )
+
+    @classmethod
+    def Generic_M42_Spacer_0_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_1mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_15mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_15mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_2mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_25mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_25mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_35mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_35mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_6mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_6mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_75mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_75mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_0_8mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_0_8mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_1mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_15mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_15mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_2mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_25mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_25mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_35mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_35mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_6mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_6mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_75mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_75mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_0_8mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_0_8mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_1mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_15mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_15mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_2mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_25mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_25mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_35mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_35mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_6mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_6mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_75mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_75mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_0_8mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_0_8mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_0_75mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_0_75mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M72_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M72_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M72_Spacer_0_75mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M72_Spacer_0_75mm"])
+
+    @classmethod
+    def Generic_M72_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M72_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M82_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M82_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M82_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M82_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M82_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M82_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M84_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M84_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M92_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M92_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M117_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M117_Spacer_1mm"])
+
+    @classmethod
+    def Generic_M117_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M117_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_1_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_1_1mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_1_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_1_3mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_1_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_1_7mm"])
+
+    @classmethod
+    def Generic_M42_Spacer_1_9mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M42_Spacer_1_9mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_1_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_1_1mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_1_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_1_3mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_1_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_1_7mm"])
+
+    @classmethod
+    def Generic_M48_Spacer_1_9mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M48_Spacer_1_9mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_1_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_1_1mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_1_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_1_3mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_1_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_1_7mm"])
+
+    @classmethod
+    def Generic_M54_Spacer_1_9mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M54_Spacer_1_9mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1_1mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1_3mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1_3mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1_7mm"])
+
+    @classmethod
+    def Generic_M68_Spacer_1_9mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M68_Spacer_1_9mm"])
+
+    @classmethod
+    def Generic_M56_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M56_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M56_Spacer_0_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M56_Spacer_0_7mm"])
+
+    @classmethod
+    def Generic_M56_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M56_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M63_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M63_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M63_Spacer_0_7mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M63_Spacer_0_7mm"])
+
+    @classmethod
+    def Generic_M63_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M63_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_SC_Schmidt_Cassegrain_Spacer_0_5mm(cls):
+        return cls.from_database(
+            cls._DATABASE["Generic_SC_Schmidt_Cassegrain_Spacer_0_5mm"]
+        )
+
+    @classmethod
+    def Generic_SC_Schmidt_Cassegrain_Spacer_1_5mm(cls):
+        return cls.from_database(
+            cls._DATABASE["Generic_SC_Schmidt_Cassegrain_Spacer_1_5mm"]
+        )
+
+    @classmethod
+    def Generic_1_25_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_1_25_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_1_25_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_1_25_Spacer_1mm"])
+
+    @classmethod
+    def Generic_1_25_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_1_25_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_1_25_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_1_25_Spacer_2mm"])
+
+    @classmethod
+    def Generic_2_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_2_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_2_Spacer_1mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_2_Spacer_1mm"])
+
+    @classmethod
+    def Generic_2_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_2_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_2_Spacer_2mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_2_Spacer_2mm"])
+
+    @classmethod
+    def Generic_M92_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M92_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M92_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M92_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M117_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M117_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M117_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M117_Spacer_1_5mm"])
+
+    @classmethod
+    def Generic_M84_Spacer_0_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M84_Spacer_0_5mm"])
+
+    @classmethod
+    def Generic_M84_Spacer_1_5mm(cls):
+        return cls.from_database(cls._DATABASE["Generic_M84_Spacer_1_5mm"])
